style(tfhe): rename module
